feat: Implementación de un nuevo handler para eliminar mensajes. nueva propiedad "seen" añadida a la entidad Message. nuevo metodo en el repositorio de mensajes que permite la busqueda de mensajes por usuario y no vistos. implementacion de un metodo para obtener mensajes no vistos. se añadio la nueva propiedad: UnreadMessageList a messages.proto. correccion al database manager la id de los mensajes paso de ser Long a Integer. nuevo metodo en el repositorio que verifica si existe un mensaje por su id. modificacion a messageService: ahora el constructor acepta otro parametro: ContactService. processReceivedMessage ahora verifica si el wsMessage contiene una lista de mensajes sin leer, valida si los mensajes ya existen en la db local para evitar duplicados. nuevo metodo processUnreadMessages que funciona como metodo auxiliar en la nueva validacion.

+++ websocket-client/src/main/java/com/pola/database/DatabaseManager.java
@@ -104,7 +104,7 @@ public class DatabaseManager {
             // Tabla de mensajes
             String createMessagesTable = """
                 CREATE TABLE IF NOT EXISTS messages (
-                    id LONG PRIMARY KEY,
+                    id INTEGER PRIMARY KEY,
                     contact_username TEXT NOT NULL,
                     content TEXT NOT NULL,
                     sender_id TEXT NOT NULL,
